requirements.txt
click==7.1.2
Pillow==7.2.0
